Update IC_BrivGemFarm_Functions.ahk
*Experimental* change to improve reliability/recovery if memory reads go wrong.
Previously, if the SB stack count cannot be read correctly, offline stacking will enter an infinite loop. There was a limit of 19 retries maximum, but it didn't work.
This change reduces the number of allowed retries to 2 - which may be okay for a lot of users, and possibly totally NOT okay for others.

Consider this commit more as "food for thought", not as a proposed change.