fix multipart content-length for saves
HTTP packets are required to have a CRLF end-of-line marker.
WinHTTP does auto-correct this, but does not update the 
content-length for multipart/form-data packets. 
That caused to send a malformed package which is rejected on
some servers.

See https://www.rfc-editor.org/rfc/rfc2616#section-2.2

Adding the CR explicitly to the existing LF does fix the 
`Content-Length` of the packet.