Refactor to avoid g_SF.Memory.ReadGemsSpent()
Some of us have gotten capped Gems Spent stats and so the Stats addon can no longer have accurate gems stats if we buy chests.

This is just an example workaround I've thrown together where the script keeps track of the gems it has spent on chests itself.

It's a bit hacky - so I don't know if this is how you would want it implemented - or if you'll maybe want to edit it so it only does this when ReadGemsSpent() is at capped Int?