Adding support for Briv's Thunder Step feat.
Thunder Step increases Briv's earned stacks during conversion by 1.2x (rounded down) - but this does not happen when the script restarts.

So - adding a way of checking if Briv has the feat equipped and if so - doing the multiplication so that `CallPreventStackFail` will apply it. Otherwise it risks a doom-loop.

Adding some MemoryFunctions methods:
- HeroHasFeatSavedInFormation
- HeroHasAnyFeatsSavedInFormation
- GetHeroFeats

And adding a method that uses them to BrivGemFarm_Performance:
- BrivHasThunderStep