Added a sleep to the catch block in the verify method.
Timing error: An invalid memory read critical error can occur when accessing a class object via com while the script is in a try block and does not have enough time to perform the read before the catch block is activated.
Putting the sleep inside the catch block appears to mitigate if not prevent the failure on the catch block activation.